backend_flask/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
passlib==1.7.4
psycopg2-binary==2.9.9
